add conftest.py for centralized test; add fixture to test_dancers.py
